+++ proxy-service/src/main/java/com/szepep/dixa/primes/proxy/PrimeService.java
@@ -2,7 +2,13 @@ package com.szepep.dixa.primes.proxy;
 
 import reactor.core.publisher.Flux;
 
+/**
+ * Service for generating prime numbers
+ */
 public interface PrimeService {
 
-    Flux<Long> prime(Long number);
+    /**
+     * Returns a flux of prime numbers
+     */
+    Flux<Integer> prime(int number);
 }
